+++ backend/src/main/java/com/minibeit/mail/service/dto/condition/ApproveCancelPostCondition.java
@@ -1,4 +1,4 @@
-package com.minibeit.mail.condition;
+package com.minibeit.mail.service.dto.condition;
 
 import org.thymeleaf.TemplateEngine;
 import org.thymeleaf.context.Context;
@@ -6,11 +6,11 @@ import org.thymeleaf.context.Context;
 import javax.mail.MessagingException;
 import javax.mail.internet.MimeMessage;
 
-public class ApplicantCancelPostCondition implements MailPostCondition {
+public class ApproveCancelPostCondition implements MailPostCondition {
     @Override
     public <T> MimeMessage makeMimeMessage(MimeMessage mimeMailMessage, TemplateEngine templateEngine, String toEmail, T data) throws MessagingException {
         mimeMailMessage.addRecipients(MimeMessage.RecipientType.TO, toEmail);
-        mimeMailMessage.setSubject("[미니바이트] 고객님의 모집 일정 내 참여확정자가 신청을 취소하여 안내해드립니다.");
+        mimeMailMessage.setSubject("[미니바이트] 고객님의 확정되었던 참여 일정이 취소되어 안내해드립니다.");
         mimeMailMessage.setText(setContext(templateEngine, data), "utf-8", "html");
         return mimeMailMessage;
     }
@@ -18,6 +18,6 @@ public class ApplicantCancelPostCondition implements MailPostCondition {
     private <T> String setContext(TemplateEngine templateEngine, T data) {
         Context context = new Context();
         context.setVariable("result", data);
-        return templateEngine.process("applicantCancel", context);
+        return templateEngine.process("approveCancel", context);
     }
 }
