+++ backend/src/main/java/com/minibeit/mail/service/dto/condition/MailCondition.java
@@ -1,4 +1,4 @@
-package com.minibeit.mail.condition;
+package com.minibeit.mail.service.dto.condition;
 
 public enum MailCondition {
     APPROVE, REJECT, APPROVECANCEL, APPLICANTCANCEL, VERIFICATION
