+++ backend/src/main/java/com/minibeit/mail/service/dto/condition/RejectPostCondition.java
@@ -1,4 +1,4 @@
-package com.minibeit.mail.condition;
+package com.minibeit.mail.service.dto.condition;
 
 import org.thymeleaf.TemplateEngine;
 import org.thymeleaf.context.Context;
@@ -6,18 +6,18 @@ import org.thymeleaf.context.Context;
 import javax.mail.MessagingException;
 import javax.mail.internet.MimeMessage;
 
-public class ApproveCancelPostCondition implements MailPostCondition {
+public class RejectPostCondition implements MailPostCondition {
     @Override
     public <T> MimeMessage makeMimeMessage(MimeMessage mimeMailMessage, TemplateEngine templateEngine, String toEmail, T data) throws MessagingException {
         mimeMailMessage.addRecipients(MimeMessage.RecipientType.TO, toEmail);
-        mimeMailMessage.setSubject("[미니바이트] 고객님의 확정되었던 참여 일정이 취소되어 안내해드립니다.");
+        mimeMailMessage.setSubject("[미니바이트] 신청하신 참여 일정의 반려를 안내해드립니다.");
         mimeMailMessage.setText(setContext(templateEngine, data), "utf-8", "html");
         return mimeMailMessage;
     }
 
-    private <T> String setContext(TemplateEngine templateEngine, T data) {
+    private String setContext(TemplateEngine templateEngine, Object data) {
         Context context = new Context();
         context.setVariable("result", data);
-        return templateEngine.process("approveCancel", context);
+        return templateEngine.process("applicantReject", context);
     }
 }
