+++ backend/src/main/java/com/minibeit/message/service/SmsService.java
@@ -2,9 +2,10 @@ package com.minibeit.message.service;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.minibeit.message.dto.MessageDto;
-import com.minibeit.message.dto.SmsMessageRequest;
-import com.minibeit.message.dto.SmsResponse;
+import com.minibeit.message.service.component.SmsProps;
+import com.minibeit.message.service.dto.MessageDto;
+import com.minibeit.message.service.dto.SmsMessageRequest;
+import com.minibeit.message.service.dto.SmsResponse;
 import com.minibeit.user.domain.User;
 import com.minibeit.user.domain.UserVerificationCode;
 import com.minibeit.user.domain.VerificationKinds;
@@ -13,7 +14,6 @@ import com.minibeit.user.domain.repository.UserVerificationCodeRepository;
 import com.minibeit.user.service.exception.UserNotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.apache.commons.codec.binary.Base64;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
@@ -36,14 +36,7 @@ import java.util.Optional;
 @Transactional
 @RequiredArgsConstructor
 public class SmsService {
-    @Value("${sms.serviceId}")
-    private String serviceId;
-    @Value("${sms.accessKey}")
-    private String accessKey;
-    @Value("${sms.secretKey}")
-    private String secretKey;
-    @Value("${sms.from}")
-    private String from;
+    private final SmsProps smsProps;
     private final UserRepository userRepository;
     private final UserVerificationCodeRepository userVerificationCodeRepository;
 
@@ -62,7 +55,7 @@ public class SmsService {
         String content = "[미니바이트] 인증번호 [" + userVerificationCode.getCode() + "] 를 입력해주세요.";
         List<MessageDto> messageDtoList = MessageDto.build(receiverPhoneNumber, content);
 
-        SmsMessageRequest smsMessageRequest = SmsMessageRequest.makeSmsRequest(from, messageDtoList);
+        SmsMessageRequest smsMessageRequest = SmsMessageRequest.makeSmsRequest(smsProps.getFrom(), messageDtoList);
 
         ObjectMapper objectMapper = new ObjectMapper();
         String jsonBody = objectMapper.writeValueAsString(smsMessageRequest);
@@ -70,7 +63,7 @@ public class SmsService {
         HttpHeaders headers = new HttpHeaders();
         headers.setContentType(MediaType.APPLICATION_JSON);
         headers.set("x-ncp-apigw-timestamp", time.toString());
-        headers.set("x-ncp-iam-access-key", this.accessKey);
+        headers.set("x-ncp-iam-access-key", smsProps.getAccessKey());
         String sig = makeSignature(time);
         headers.set("x-ncp-apigw-signature-v2", sig);
 
@@ -79,17 +72,17 @@ public class SmsService {
         RestTemplate restTemplate = new RestTemplate();
         restTemplate.setRequestFactory(new HttpComponentsClientHttpRequestFactory());
 
-        return restTemplate.postForObject(new URI("https://sens.apigw.ntruss.com/sms/v2/services/" + this.serviceId + "/messages"), body, SmsResponse.class);
+        return restTemplate.postForObject(new URI("https://sens.apigw.ntruss.com/sms/v2/services/" + smsProps.getServiceId() + "/messages"), body, SmsResponse.class);
     }
 
     private String makeSignature(Long time) throws NoSuchAlgorithmException, InvalidKeyException {
         String space = " ";
         String newLine = "\n";
         String method = "POST";
-        String url = "/sms/v2/services/" + this.serviceId + "/messages";
+        String url = "/sms/v2/services/" + smsProps.getServiceId() + "/messages";
         String timestamp = time.toString();
-        String accessKey = this.accessKey;
-        String secretKey = this.secretKey;
+        String accessKey = smsProps.getAccessKey();
+        String secretKey = smsProps.getSecretKey();
 
         String message = new StringBuilder()
                 .append(method)
