+++ backend/src/test/java/com/minibeit/message/web/SmsControllerTest.java
@@ -1,8 +1,8 @@
 package com.minibeit.message.web;
 
 import com.minibeit.MvcTest;
-import com.minibeit.message.dto.SmsRequest;
-import com.minibeit.message.dto.SmsResponse;
+import com.minibeit.message.service.dto.SmsRequest;
+import com.minibeit.message.service.dto.SmsResponse;
 import com.minibeit.message.service.SmsService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
